% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_construction.R
\name{construct_empirical_pmf_log_df}
\alias{construct_empirical_pmf_log_df}
\title{Construct Empirical PMF with Log Probabilities}
\usage{
construct_empirical_pmf_log_df(simulation_result, n_prey_initial, alpha)
}
\arguments{
\item{simulation_result}{A dataframe containing simulation results with a column
\code{n_prey_remaining} indicating the number of prey remaining.}

\item{n_prey_initial}{An integer representing the initial number of prey. It defines
the range of possible values for \code{n_prey_remaining}.}

\item{alpha}{A numeric value representing the hyperparameter for the Dirichlet Process prior.
It controls the influence of the prior in the posterior calculation.}
}
\value{
A tibble with the following columns:
\item{n_prey_remaining}{An integer indicating the number of prey remaining.}
\item{pmf}{The empirical probability mass function for each value of \code{n_prey_remaining}.}
\item{log_prob}{The logarithm of the PMF.}
}
\description{
This function constructs a dataframe containing the empirical probability mass function (PMF)
and its logarithm from simulation results. It computes the PMF using a Dirichlet Process
posterior mean, with a uniform prior over all possible values of \code{n_prey_remaining}. The
resulting dataframe includes the PMF and its logarithm for each possible value of \code{n_prey_remaining}.
}
\details{
The PMF is computed using the Dirichlet Process posterior mean, where the prior measure is
uniform across all possible values of \code{n_prey_remaining}. The PMF is adjusted by the hyperparameter
\code{alpha}, and its logarithm is calculated for each value.
}
\examples{
# Example usage with hypothetical simulation results and parameters
simulation_result <- tibble::tibble(n_prey_remaining = c(0, 1, 1, 2, 2, 2, 3))
n_prey_initial <- 3
alpha <- 1

# Construct the empirical PMF with log probabilities
df_pmf_log <- frbayes:::construct_empirical_pmf_log_df(simulation_result, n_prey_initial, alpha)
print(df_pmf_log)

}
