% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_construction.R
\name{create_approx_counts}
\alias{create_approx_counts}
\title{Create Approximate Counts from Simulation Results}
\usage{
create_approx_counts(simulation_result, n_prey_initial)
}
\arguments{
\item{simulation_result}{A dataframe containing simulation results with a column
\code{n_prey_remaining} indicating the number of prey remaining.}

\item{n_prey_initial}{An integer representing the initial number of prey. It defines
the range of possible values for \code{n_prey_remaining}.}
}
\value{
A tibble with two columns:
\item{n_prey_remaining}{An integer indicating the number of prey remaining.}
\item{n}{An integer indicating the count of occurrences for each value of
\code{n_prey_remaining}. Missing values are replaced with zero.}
}
\description{
This function processes a dataframe of simulation results to count occurrences of each
value of \code{n_prey_remaining}. It ensures that all possible values of \code{n_prey_remaining}
from 0 to \code{n_prey_initial} are represented, even if they were not observed in the
simulation. It replaces missing values with zero and provides a complete dataframe
with counts for each possible value.
}
\examples{
# Example simulation result
simulation_result <- tibble::tibble(n_prey_remaining = c(0, 1, 1, 2, 2, 2, 3))

# Generate approximate counts
frbayes:::create_approx_counts(simulation_result, n_prey_initial = 3)

}
