% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_generalised_holling}
\alias{model_generalised_holling}
\title{Generate a Generalized Holling Model Function for Prey-Predator Simulation}
\usage{
model_generalised_holling()
}
\value{
A function that calculates the propensity of prey consumption based on the current
number of prey and a set of parameters. The returned function takes two arguments:
\describe{
\item{prey}{Numeric representing the current number of prey.}
\item{parameters}{A list containing the parameters for the model. Must include:}
\describe{
\item{b}{A numeric value representing the maximum rate of consumption.}
\item{h}{A numeric value representing the handling time.}
\item{q}{A numeric value representing the type of functional response.}
}
}
}
\description{
This function returns a model function based on the Generalized Holling model, which describes
the rate of prey consumption as a function of the number of prey and three parameters:
\code{b} (the maximum rate of consumption), \code{h} (the handling time), and \code{q} (the type of functional response).
}
\details{
The Generalized Holling model is defined as:
\deqn{ \text{rate} = \frac{b \cdot \text{prey}^{(1 + q)}}{1 + b \cdot h \cdot \text{prey}^{(1 + q)} }}
where \code{b} is the maximum rate of consumption, \code{h} is the handling time, and \code{q}
modifies the type of functional response.
}
\examples{
model <- model_generalised_holling()
parameters <- list(b = 0.5, h = 0.2, q = 1)
model(10, parameters) # Computes the rate of prey consumption for 10 prey

}
