% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_stochastic_degradation}
\alias{model_stochastic_degradation}
\title{Generate a Constant Rate Model for Prey-Predator Simulation}
\usage{
model_stochastic_degradation()
}
\value{
A function that computes the propensity of prey consumption based on the current
number of prey and a set of parameters. The returned function takes two arguments:
\describe{
\item{prey}{Integer representing the current number of prey.}
\item{parameters}{A list containing the parameters for the model. Must include:}
\describe{
\item{rate}{A numeric value representing the constant rate of prey consumption per prey.}
}
}
}
\description{
This function returns a model function where the rate of prey consumption
is constant and proportional to the current number of prey. The returned function
can be used in simulations of prey-predator interactions.
}
\examples{
model <- model_stochastic_degradation()
parameters <- list(rate = 0.1)
model(10, parameters) # Returns 1.0

}
