% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_many_trajectories}
\alias{simulate_many_trajectories}
\title{Simulate Multiple Runs of a Prey-Predator Model}
\usage{
simulate_many_trajectories(
  n_trajectories,
  n_prey_initial,
  time_max,
  model,
  parameters
)
}
\arguments{
\item{n_trajectories}{An integer representing the number of trajectories to simulate.}

\item{n_prey_initial}{An integer representing the initial number of prey for each trajectory.}

\item{time_max}{A numeric value indicating the maximum time for the simulations.}

\item{model}{A function that calculates the propensity given the current number of prey and parameters.}

\item{parameters}{A list of named parameters required by the model function.}
}
\value{
A data frame with columns \code{time}, \code{n_prey_remaining}, \code{n_prey_eaten},
and \code{trajectory_id}, representing the simulated data for all trajectories.
}
\description{
This function simulates multiple runs of a prey-predator model over a specified
period of time, given initial conditions and model parameters. The simulation
algorithm used is the Gillespie algorithm.
}
