% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_single}
\alias{simulate_single}
\title{Simulate a Single Run of a Prey-Predator Model}
\usage{
simulate_single(n_prey_initial, time_max, model, parameters)
}
\arguments{
\item{n_prey_initial}{An integer representing the initial number of prey.}

\item{time_max}{A numeric value indicating the maximum time for the simulation.}

\item{model}{A function that calculates the propensity given the current number of prey and parameters.}

\item{parameters}{A list of named parameters required by the model function.}
}
\value{
The number of prey remaining at the end of the simulation.
}
\description{
This function simulates a single run of a prey-predator model over a specified
period of time, given initial conditions and model parameters.
}
