% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_trajectory}
\alias{simulate_trajectory}
\title{Simulate a Single Run of a Prey-Predator Model}
\usage{
simulate_trajectory(n_prey_initial, time_max, model, parameters)
}
\arguments{
\item{n_prey_initial}{An integer representing the initial number of prey.}

\item{time_max}{A numeric value indicating the maximum time for the simulation.}

\item{model}{A function that calculates the propensity given the current number of prey and parameters.}

\item{parameters}{A list of named parameters required by the model function.}
}
\value{
A data frame with numbers of prey remaining (and eaten) by time.
}
\description{
This function simulates a single run of a prey-predator model over a specified
period of time, given initial conditions and model parameters. The simulation
algorithm used is the Gillespie algorithm.
}
